// web/src/api/generated/model/adminModelResponse.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * matthew-api
 * OpenAPI spec version: 0.1.0
 */
import type { AdminModelResponseStatus } from './adminModelResponseStatus';

export interface AdminModelResponse {
  id: number;
  name: string;
  context_window: number;
  max_output_tokens: number;
  status: AdminModelResponseStatus;
}